Move `util::Response` into a submodule
